+++ backend/src/lambda/http/updateTodo.ts
@@ -4,7 +4,7 @@ import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda'
 import * as middy from 'middy'
 import { cors, httpErrorHandler } from 'middy/middlewares'
 
-import { updateTodoItem, todoItemExists } from '../../helpers/todos'
+import { updateTodoItem, todoItemExists } from '../../businessLogic/todos'
 import { UpdateTodoRequest } from '../../requests/UpdateTodoRequest'
 
 export const handler = middy(
